Switch to a portability service hosted in AWS API Gateway
This commit removes the old flask app and replaces it with an serverless
API using AWS's API Gateway and Lambda. It uses Swagger to define a
first draft of the API to the portability service.

The service currently works with the local Cromwell execution
environment defined in another environment. It's also a core component
of the HCA's contribution to the GA4GH's workflow interoperability
testbed.